test: register render loss contract
