+++ resources/normative-clauses.txt
@@ -180,6 +180,7 @@
 1  R1b A MATCHER'S COORDINATES ARE LOCAL, NOT ABSOLUTE
 1  RAW PASSTHROUGH
 1  RENDER ORDER
+1  RENDER-LOSS REPORTING
 1  RESOLVED OR NOT
 1  SECURITY REQUIREMENTS
 1  SEMANTIC SPAN ATTRIBUTES
